tests/wpt/web-platform-tests/fetch/api/request/request-disturbed.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Request disturbed</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#request">
    <meta name="help" href="https://fetch.spec.whatwg.org/#body-mixin">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      var initValuesDict = {"method" : "POST",
                            "body" : "Request's body"
      };

      var noBodyConsumed = new Request("");
      noBodyConsumed.blob();
      var bodyConsumed = new Request("", initValuesDict);
      bodyConsumed.blob();

      test(function() {
        assert_false(noBodyConsumed.bodyUsed , "bodyUsed is false when request is not disturbed");
        try {
          noBodyConsumed.clone();
        } catch (e) {
          assert_unreached("Can use request not disturbed for creating or cloning request");
        }
      }, "Request without body cannot be disturbed");

      test(function() {
        assert_true(bodyConsumed.bodyUsed , "bodyUsed is true when request is disturbed");
        assert_throws(new TypeError(), function() { bodyConsumed.clone(); });
      }, "Check cloning a disturbed request");

      test(function() {
        assert_true(bodyConsumed.bodyUsed , "bodyUsed is true when request is disturbed");
        assert_throws(new TypeError(), function() { new Request(bodyConsumed); });
      }, "Check creating a new request from a disturbed request");

      test(function() {
        var bodyRequest = new Request("", initValuesDict);
        assert_false(bodyRequest.bodyUsed , "bodyUsed is false when request is not disturbed");
        var requestFromRequest = new Request(bodyRequest);
        assert_true(bodyRequest.bodyUsed , "bodyUsed is true when request is disturbed");
      }, "Input request used for creating new request became disturbed");

      promise_test(function(test) {
        assert_true(bodyConsumed.bodyUsed , "bodyUsed is true when request is disturbed");
        return promise_rejects(test, new TypeError(), bodyConsumed.blob());
      }, "Check consuming a disturbed request");
    </script>
  </body>
</html>
